backend/requirements.txt
crewai[litellm]
crewai-tools
fastapi
uvicorn[standard]
tavily-python
pyyaml
python-dotenv
pydantic
pandas
supabase
